Pepper/pepper_DL/client/requirements2.txt
numpy
pandas
matplotlib
Flask
Pillow
opencv-python
requests
tqdm
PyYAML
seaborn
scikit-image
filterpy
cython-bbox
# Need to install a compatible version of PyTorch. I did not include this because it can vary from machine to machine.